refactor move QueryStatus to common
