+++ tests/integration_tests/cache_tests.py
@@ -20,8 +20,8 @@ import json
 import pytest
 
 from superset import app, db
+from superset.common.db_query_status import QueryStatus
 from superset.extensions import cache_manager
-from superset.utils.core import QueryStatus
 from tests.integration_tests.fixtures.birth_names_dashboard import (
     load_birth_names_dashboard_with_slices,
 )
