+++ tests/integration_tests/queries/api_tests.py
@@ -27,8 +27,9 @@ from sqlalchemy.sql import func
 
 import tests.integration_tests.test_app
 from superset import db, security_manager
+from superset.common.db_query_status import QueryStatus
 from superset.models.core import Database
-from superset.utils.core import get_example_database, get_main_database, QueryStatus
+from superset.utils.core import get_example_database, get_main_database
 from superset.models.sql_lab import Query
 
 from tests.integration_tests.base_tests import SupersetTestCase
